Add checkbox to the post-orcid-profile thingie
