+++ keycloak/keycloakify/src/login/pages/TermsAcceptance.tsx
@@ -1,0 +1,55 @@
+import { kcSanitize } from "keycloakify/lib/kcSanitize";
+import type { KcClsx } from "keycloakify/login/lib/kcClsx";
+import type { KcContext } from "../KcContext";
+import type { I18n } from "../i18n";
+
+export function TermsAcceptance(props: {
+    i18n: I18n;
+    kcClsx: KcClsx;
+    messagesPerField: Pick<KcContext["messagesPerField"], "existsError" | "get">;
+    areTermsAccepted: boolean;
+    onAreTermsAcceptedValueChange: (areTermsAccepted: boolean) => void;
+}) {
+    const { i18n, kcClsx, messagesPerField, areTermsAccepted, onAreTermsAcceptedValueChange } = props;
+
+    const { msg } = i18n;
+
+    return (
+        <>
+            <div className="form-group">
+                <div className={kcClsx("kcInputWrapperClass")}>
+                    {msg("termsTitle")}
+                    <div id="kc-registration-terms-text">{msg("termsText")}</div>
+                </div>
+            </div>
+            <div className="form-group">
+                <div className={kcClsx("kcLabelWrapperClass")}>
+                    <input
+                        type="checkbox"
+                        id="termsAccepted"
+                        name="termsAccepted"
+                        className={kcClsx("kcCheckboxInputClass")}
+                        checked={areTermsAccepted}
+                        onChange={e => onAreTermsAcceptedValueChange(e.target.checked)}
+                        aria-invalid={messagesPerField.existsError("termsAccepted")}
+                    />
+                    <label htmlFor="termsAccepted" className={kcClsx("kcLabelClass")}>
+                        {msg("acceptTerms")}
+                    </label>
+                </div>
+                {messagesPerField.existsError("termsAccepted") && (
+                    <div className={kcClsx("kcLabelWrapperClass")}>
+                        <span
+                            id="input-error-terms-accepted"
+                            className={kcClsx("kcInputErrorMessageClass")}
+                            aria-live="polite"
+                            dangerouslySetInnerHTML={{
+                                __html: kcSanitize(messagesPerField.get("termsAccepted"))
+                            }}
+                        />
+                    </div>
+                )}
+            </div>
+        </>
+    );
+}
